Updates Dockerfile for integration testing
In order to incorporate the `remote_val_harness` into the KMS' testing
cycle, we need to use a multi-stage Docker build. This allows us to grab
specific harness versions from the Tendermint `remote_val_harness`
Docker image without influencing the underlying layers of the final
build image (which will still be the same centos:7-based build as
previously, but now with the `remote_val_harness` binary and
configuration included).